requirements.txt
aws-cdk-lib==2.37.1
aws-cdk.aws-lambda-python-alpha==2.37.1a0
constructs==10.1.72
